wallet&card

+++ src/main/java/com/group/teona/repositories/AdressRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdressRepository extends JpaRepository<Adress,Integer> {
+public interface AdressRepository extends JpaRepository<Adress,Long> {
 }
